requirements.txt
mechanicalsoup
python-nmap 
termcolor
